Finish first section of landing page
While the first section is complete, I did stumble a bit in the code and need to clean it up so it's
better organized and easy to understand if I come back after spending some time away.

Toughest part was not knowing if the header and hero text are too centered and should be pushed more to the side, though since
i'm using a 1440p monitor the result may seem weirder to me and it's actually fine. Need to look into some way ensure that a website follows proper screen size proportions.