Finish all sections and website
I have finished all the remaining sections as well as changed the text to fit the "theme" of the website. I called it CURSEMPORIUM and it's a website that sells cursed items.
While this is technically the end of the project as I have done what was required, I still plan on making something of a version 1.2 of sorts, basically just going to clean up
the code and better organize it. I ended up changing some classes'names to better fitting divs but that leaves some with odd names.
I can also probably still reduce the code by overhauling the css sheet for the next version as well.

Overall, very fun and relatively easy to do, I can see how html and css are fun to work in and looking online, there is no shortage of tutorials for implementing cool features. Might add a snap-scroll option in the next version as well.
But I will also continue onwards in The Odin Project. Introduction to Javascript is next and I'm pretty excited for it.